+++ client_src/js/exceptions/octofarm-client.exceptions.js
@@ -1,11 +1,19 @@
 import { errorTypes } from "./error.types";
 
+//TODO: Needs updating for actual usage.
 export const ClientErrors = {
-  FAILED_VALIDATION: {
+  FAILED_VALIDATION_PATH: {
     type: errorTypes.CLIENT,
     color: "danger",
     code: "FAILED_VALIDATION_PATH",
     message: "Input validation failed, no path was supplied",
+    statusCode: 412
+  },
+  FAILED_VALIDATION_KEY: {
+    type: errorTypes.CLIENT,
+    color: "danger",
+    code: "FAILED_VALIDATION_KEY",
+    message: "Input validation failed, no key was supplied",
     statusCode: 412
   },
   FAILED_STATE_UPDATE: {