+++ client_src/js/lib/functions/ui.js
@@ -108,14 +108,19 @@ export default class UI {
       });
     });
   }
+  //TODO: Move to a templates folder
+  static returnSpinnerTemplate() {
+    return "<i class='fas fa-spinner fa-spin'></i>";
+  }
+
   static removeLoaderFromElementInnerHTML(element) {
     if (element.innerHTML.includes("spinner")) {
-      element.innerHTML = element.innerHTML.replace('<i class="fas fa-spinner fa-spin"></i>', "");
+      element.innerHTML = element.innerHTML.replace(UI.returnSpinnerTemplate(), "");
     }
   }
   static addLoaderToElementsInnerHTML(element) {
     if (!element.innerHTML.includes("spinner")) {
-      element.innerHTML += ' <i class="fas fa-spinner fa-spin"></i>';
+      element.innerHTML += UI.returnSpinnerTemplate();
     }
   }
   static checkIfAnyModalShown() {