+++ client_src/js/monitoring/monitoring.updater.js
@@ -9,7 +9,7 @@ import {
 } from "../lib/modules/Printers/actionButtons.js";
 import OctoPrintClient from "../lib/octoprint.js";
 import { checkTemps } from "../lib/modules/temperatureCheck.js";
-import { checkFilamentManager } from "../lib/modules/filamentGrab.js";
+import { checkFilamentManager } from "../services/filament-manager-plugin.service";
 import doubleClickFullScreen from "../lib/functions/fullscreen.js";
 import OctoFarmClient from "../services/octofarm-client.service";
 import { getControlList, getPrinterInfo } from "./monitoring-view.state";